+++ src/storm-dft/modelchecker/dft/DFTModelChecker.cpp
@@ -424,10 +424,7 @@ namespace storm {
                     storm::api::exportSparseModelAsDrn(model, ioSettings.getExportExplicitFilename(), parameterNames);
                 }
                 if (ioSettings.isExportDotSet()) {
-                    std::ofstream stream;
-                    storm::utility::openFile(ioSettings.getExportDotFilename(), stream);
-                    model->writeDotToStream(stream, true, true);
-                    storm::utility::closeFile(stream);
+                    storm::api::exportSparseModelAsDot(model, ioSettings.getExportDotFilename(), ioSettings.getExportDotMaxWidth());
                 }
 
                 // Model checking