+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Version 1.3.x
 - Support for export of MTBDDs from Storm.
 - New settings module `transformation` for Markov chain transformations. Use `--help transformation` to get a list of available transformations.
 - Support for eliminating chains of Non-Markovian states in MAs via `--eliminate-chains`.
+- Export to dot format allows for maximal line width in states (argument `--dot-maxwidth <width>`)
 - `storm-conv` can now apply transformations on a prism file.
 - `storm-dft`: Support partial-order for state space generation.
 - `storm-dft`: Compute lower and upper bounds for number of BE failures via SMT.